Fixed gluemnc and xplodemnc (finally!!).  Now works with multiple tiles in X and Y directions, as well as with float and double variable precision.
It's not terribly fast, but allows combination of all *t0??.nc tiles to a single *.glob.nc file.  Works with dynamics output and tave output as well.  I don't know if it works for pickup files, but it may.

It should be possible to -omp multi-thread it by changing the -DEBUG flag to include the omp directives if your NCO operators were compiled with omp support.

Happy gluing!


Former-commit-id: ac4d50f61fa07e5e2610f8ee320d1a9b41f1ea00